+++ test_allianceutils/tests/profile_auth/urls.py
@@ -1,11 +1,9 @@
-from django.conf.urls import include
-from django.conf.urls import url
+from django.urls import path
 
 from .views import LoginRequiredView
 
 app_name = 'profile_auth'
 
 urlpatterns = [
-    url(r'^accounts/', include('authtools.urls')),
-    url(r'^protected/$', LoginRequiredView.as_view(), name='login_required')
+    path(r'protected/', LoginRequiredView.as_view(), name='login_required')
 ]