+++ test_allianceutils/urls.py
@@ -8,9 +8,9 @@ import test_allianceutils.tests.profile_auth.urls
 import test_allianceutils.tests.viewset_permissions.urls
 
 urlpatterns = [
-    path(r'accounts/login/', LoginView.as_view(), name='login'),
-    path(r'accounts/logout/', LogoutView.as_view(), name='logout'),
-    path(r'middleware/', include(test_allianceutils.tests.middleware.urls)),
-    path(r'profile_auth/', include(test_allianceutils.tests.profile_auth.urls)),
-    path(r'viewset_permissions/', include(test_allianceutils.tests.viewset_permissions.urls)),
+    path('accounts/login/', LoginView.as_view(), name='login'),
+    path('accounts/logout/', LogoutView.as_view(), name='logout'),
+    path('middleware/', include(test_allianceutils.tests.middleware.urls)),
+    path('profile_auth/', include(test_allianceutils.tests.profile_auth.urls)),
+    path('viewset_permissions/', include(test_allianceutils.tests.viewset_permissions.urls)),
 ]